ENH: relocate zero_one to pTraits, allows reuse for scalar clamping
ENH: add pTraits and IO for std::int8_t

STYLE: cull some implicitly available includes

- pTraits.H is included by label/scalar etc
- zero.H is included by UList

STYLE: cull redundant forward declarations for Istream/Ostream

+++ applications/test/List/Test-List.C
@@ -81,7 +81,7 @@ namespace ListPolicy
 {
 
 // Override on a per-type basis
-template<> struct short_length<short> : std::integral_constant<short,20> {};
+template<> struct short_length<short> : std::integral_constant<int,20> {};
 
 } // End namespace ListPolicy
 } // End namespace Detail
